+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/utils/SnapshotManager.java
@@ -107,7 +107,7 @@ public class SnapshotManager {
                         // Download snapshot.
                         Path snapshotZip =
                                 this.ctiClient.download(consumerInfo.getLastSnapshotLink(), this.environment);
-                        Path outputDir = this.environment.tmpFile();
+                        Path outputDir = this.environment.tmpDir();
 
                         try (DirectoryStream<Path> stream = this.getStream(outputDir)) {
                             // Unzip snapshot.