+++ plugins/content-manager/src/test/java/com/wazuh/contentmanager/ContentManagerPluginIT.java
@@ -18,8 +18,8 @@ package com.wazuh.contentmanager;
 
 import com.carrotsearch.randomizedtesting.annotations.ThreadLeakScope;
 
-import org.apache.http.ParseException;
-import org.apache.http.util.EntityUtils;
+import org.apache.hc.core5.http.ParseException;
+import org.apache.hc.core5.http.io.entity.EntityUtils;
 import org.opensearch.client.Request;
 import org.opensearch.client.Response;
 import org.opensearch.plugins.Plugin;